refactor: put all slack related classes into one package
