+++ .eslintrc.json
@@ -56,7 +56,7 @@
         "devDependencies": [
           "**/test/**",
           "**/build-tools/**",
-          "src/daily-spends/digest-slack-notification.lambda.ts",
+          "src/daily-spends/slack/digest-notification.lambda.ts",
           "src/daily-spends/yesterday.lambda.ts"
         ],
         "optionalDependencies": false,
