+++ .gitignore
@@ -55,5 +55,5 @@ junit.xml
 tsconfig.json
 !/API.md
 /assets/
-!/src/daily-spends/digest-slack-notification-function.ts
+!/src/daily-spends/slack/digest-notification-function.ts
 !/src/daily-spends/yesterday-function.ts
