+++ .prettierignore
@@ -1,3 +1,3 @@
 # ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
-src/daily-spends/digest-slack-notification-function.ts
+src/daily-spends/slack/digest-notification-function.ts
 src/daily-spends/yesterday-function.ts
