+++ package.json
@@ -9,8 +9,8 @@
     "build": "npx projen build",
     "bump": "npx projen bump",
     "bundle": "npx projen bundle",
-    "bundle:daily-spends/digest-slack-notification.lambda": "npx projen bundle:daily-spends/digest-slack-notification.lambda",
-    "bundle:daily-spends/digest-slack-notification.lambda:watch": "npx projen bundle:daily-spends/digest-slack-notification.lambda:watch",
+    "bundle:daily-spends/slack/digest-notification.lambda": "npx projen bundle:daily-spends/slack/digest-notification.lambda",
+    "bundle:daily-spends/slack/digest-notification.lambda:watch": "npx projen bundle:daily-spends/slack/digest-notification.lambda:watch",
     "bundle:daily-spends/yesterday.lambda": "npx projen bundle:daily-spends/yesterday.lambda",
     "bundle:daily-spends/yesterday.lambda:watch": "npx projen bundle:daily-spends/yesterday.lambda:watch",
     "clobber": "npx projen clobber",
