--- src/daily-spends/digest-slack-notification-function.ts
@@ -1,26 +1,0 @@
-// ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
-import * as path from 'path';
-import * as lambda from 'aws-cdk-lib/aws-lambda';
-import { Construct } from 'constructs';
-
-/**
- * Props for DigestSlackNotificationFunction
- */
-export interface DigestSlackNotificationFunctionProps extends lambda.FunctionOptions {
-}
-
-/**
- * An AWS Lambda function which executes src/daily-spends/digest-slack-notification.
- */
-export class DigestSlackNotificationFunction extends lambda.Function {
-  constructor(scope: Construct, id: string, props?: DigestSlackNotificationFunctionProps) {
-    super(scope, id, {
-      description: 'src/daily-spends/digest-slack-notification.lambda.ts',
-      ...props,
-      runtime: new lambda.Runtime('nodejs14.x', lambda.RuntimeFamily.NODEJS),
-      handler: 'index.handler',
-      code: lambda.Code.fromAsset(path.join(__dirname, '../../assets/daily-spends/digest-slack-notification.lambda')),
-    });
-    this.addEnvironment('AWS_NODEJS_CONNECTION_REUSE_ENABLED', '1', { removeInEdge: true });
-  }
-}
