+++ src/daily-spends/slack/digest-notification-function.ts
@@ -1,0 +1,26 @@
+// ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
+import * as path from 'path';
+import * as lambda from 'aws-cdk-lib/aws-lambda';
+import { Construct } from 'constructs';
+
+/**
+ * Props for DigestNotificationFunction
+ */
+export interface DigestNotificationFunctionProps extends lambda.FunctionOptions {
+}
+
+/**
+ * An AWS Lambda function which executes src/daily-spends/slack/digest-notification.
+ */
+export class DigestNotificationFunction extends lambda.Function {
+  constructor(scope: Construct, id: string, props?: DigestNotificationFunctionProps) {
+    super(scope, id, {
+      description: 'src/daily-spends/slack/digest-notification.lambda.ts',
+      ...props,
+      runtime: new lambda.Runtime('nodejs14.x', lambda.RuntimeFamily.NODEJS),
+      handler: 'index.handler',
+      code: lambda.Code.fromAsset(path.join(__dirname, '../../../assets/daily-spends/slack/digest-notification.lambda')),
+    });
+    this.addEnvironment('AWS_NODEJS_CONNECTION_REUSE_ENABLED', '1', { removeInEdge: true });
+  }
+}
