+++ src/daily-spends/slack/digest-notification.lambda.ts
@@ -4,8 +4,8 @@ import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
 import { Block } from "@slack/types";
 import { IncomingWebhook, IncomingWebhookResult } from "@slack/webhook";
 import { parse } from "csv-parse";
-import { DailySpendsDigestSlackNotification } from "./digest-slack-message";
-import { DailySpendsReport } from "./report";
+import { DailySpendsReport } from "../report";
+import { DailySpendsDigestSlackNotification } from "./digest-message";
 
 const s3Client = new S3Client({});
 const webhook = new IncomingWebhook(process.env.WEBHOOK_URL || "");
